Create ATAC_with_individual_peaks
Cells were annotated with reduced peak. Then, here we created Seurat object from individual peak and metadata from the reduced peak object is transferred to individual peak object. If there is a cell that is not present (filtered out) in the reduced object, those cells were filtered out in the final individual object.